chore: refactored directory
