--- src/com/edu/rmit/oosd/wk6/creationalpatterns/decorator/grailed/Main.java
@@ -1,4 +1,0 @@
-package com.edu.rmit.oosd.wk6.creationalpatterns.decorator.grailed;
-
-public class Main {
-}
