--- src/com/edu/rmit/oosd/wk6/creationalpatterns/decorator/starbucks/Tea.java
@@ -1,13 +1,0 @@
-package com.edu.rmit.oosd.wk6.creationalpatterns.decorator.starbucks;
-
-public class Tea extends Beverage {
-
-    public Tea() {
-        description = "Tea";
-    }
-
-    @Override
-    public double cost() {
-        return 3.50;
-    }
-}
