+++ src/com/edu/rmit/oosd/wk8/structuralpatterns/decorator/ebay/Main.java
@@ -1,0 +1,4 @@
+package com.edu.rmit.oosd.wk8.structuralpatterns.decorator.ebay;
+
+public class Main {
+}
