+++ src/com/edu/rmit/oosd/wk8/structuralpatterns/decorator/grailed/Main.java
@@ -1,0 +1,4 @@
+package com.edu.rmit.oosd.wk8.structuralpatterns.decorator.grailed;
+
+public class Main {
+}
