+++ src/com/edu/rmit/oosd/wk8/structuralpatterns/decorator/sneakervillain/decorator/abstraction/AttributeDecorator.java
@@ -1,0 +1,21 @@
+package com.edu.rmit.oosd.wk8.structuralpatterns.decorator.sneakervillain.decorator.abstraction;
+
+import com.edu.rmit.oosd.wk8.structuralpatterns.decorator.sneakervillain.component.base.Clothes;
+
+/**
+ * @author Carl Karama
+ * @apiNote Uses the decorator pattern to add attributes to a product and go in detail to customise the prices of each attribute
+ * @implSpec
+ */
+
+public abstract class AttributeDecorator extends Clothes {
+
+    protected Clothes attribute;
+
+    public String getDescription() {
+        return attribute.getDescription();
+    }
+
+    @Override
+    public abstract double cost();
+}
